api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
pydantic==2.9.0
pydantic-settings==2.6.0
httpx==0.27.2
ollama==0.3.3
celery[redis]==5.4.0
redis==5.1.0
cryptography==43.0.0
python-jose[cryptography]==3.3.0
svix==1.24.0
python-dotenv==1.0.1
structlog==24.4.0
pgvector==0.3.5
presidio-analyzer==2.2.355
presidio-anonymizer==2.2.355
spacy==3.7.4
https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.7.1/en_core_web_lg-3.7.1-py3-none-any.whl
python-multipart==0.0.9
pypdf==4.3.1
